fix(insights): move scene story play helpers into createInsightScene
@testing-library/dom and @testing-library/user-event don't resolve from products/ in the frontend typecheck, so the funnel stabilize and property-filter play functions now live in the shared harness under frontend/src and are imported by the products story file.

Generated-By: PostHog Code
Task-Id: 482e7bd9-6d6b-4b2c-8dfa-a65ba43b5a50
